Move util to core
