+++ src/contract/transition_call.rs
@@ -1,10 +1,10 @@
 use std::sync::Arc;
 
 use crate::{
+    core::parse_zil,
     core::{GetTransactionResponse, ZilAddress},
     middlewares::Middleware,
     transaction::{TransactionBuilder, TransactionParams},
-    util::parse_zil,
     Error,
 };
